--- ranx.h
@@ -1,4 +1,0 @@
-void init_ranx(long);
-double ranx();
-long get_seed_ranx();
-double gsl_box_muller();
