+++ src/ranx.h
@@ -1,0 +1,4 @@
+void init_ranx(long);
+double ranx();
+long get_seed_ranx();
+double gsl_box_muller();
